Added campt and hasblob tests
