+++ isis/tests/Fixtures.h
@@ -11,6 +11,9 @@
 
 #include <nlohmann/json.hpp>
 
+#include "csm/csm.h"
+#include "csm/Ellipsoid.h"
+
 #include "Cube.h"
 #include "IException.h"
 #include "OriginalLabel.h"
@@ -20,6 +23,8 @@
 #include "ImagePolygon.h"
 #include "PolygonTools.h"
 #include "Blob.h"
+#include "MockCsmPlugin.h"
+#include "Mocks.h"
 #include "ControlNet.h"
 #include "FileList.h"
 #include "FileName.h"
@@ -180,8 +185,8 @@ namespace Isis {
       QString ckPath = "data/mroKernels/mroCK.bc";
       QString sclkPath = "data/mroKernels/mroSCLK.tsc";
       QString lskPath = "data/mroKernels/mroLSK.tls";
-      Cube dejitteredCube; 
-      QString jitterPath; 
+      Cube dejitteredCube;
+      QString jitterPath;
 
       void SetUp() override;
       void setInstrument(QString ikid, QString instrumentId, QString spacecraftName);
@@ -208,6 +213,42 @@ namespace Isis {
 
       void SetUp() override;
   };
+
+  class CSMCubeFixture : public SmallCube {
+  protected:
+    QString filename;
+    MockRasterGM mockModel;
+
+    void SetUp() override;
+};
+
+
+class CSMCameraFixture : public CSMCubeFixture {
+  protected:
+    Camera *testCam;
+
+    void SetUp() override;
+};
+
+
+class CSMCameraSetFixture : public CSMCameraFixture {
+  protected:
+    csm::Ellipsoid wgs84;
+    csm::ImageCoord imagePt;
+    csm::EcefCoord groundPt;
+    csm::EcefLocus imageLocus;
+
+    void SetUp() override;
+};
+
+
+class CSMCameraDemFixture : public CSMCubeFixture {
+  protected:
+    Camera *testCam;
+    double demRadius;
+
+    void SetUp() override;
+};
 }
 
 #endif
