+++ isis/tests/TestUtilities.h
@@ -1,12 +1,14 @@
 #ifndef TestUtilities_h
 #define TestUtilities_h
 
-#include "gtest/gtest.h"
+#include "gmock/gmock.h"
 
 #include <string>
 #include <vector>
 
 #include <QString>
+
+#include "csm.h"
 
 #include "IException.h"
 #include "PvlGroup.h"
@@ -54,6 +56,10 @@ namespace Isis {
   void compareCsvLine(CSVReader::CSVAxis csvLine, QString headerStr, int initialIndex=0);
   void compareCsvLine(CSVReader::CSVAxis csvLine, CSVReader::CSVAxis csvLine2, int initialIndex=0,
                       double tolerance = 0.000001);
+
+  // gMock matchers for CSM structs
+  ::testing::Matcher<const csm::ImageCoord&> MatchImageCoord(const csm::ImageCoord &expected);
+  ::testing::Matcher<const csm::EcefCoord&> MatchEcefCoord(const csm::EcefCoord &expected);
 }
 
 #endif
